Pull pxtarget.json in, read target name/embedUrl from json

+++ cli/cli.ts
@@ -1917,26 +1917,44 @@ function buildWebStringsAsync() {
     return Promise.resolve()
 }
 
-function buildSkillMapAsync() {
-    return nodeutil.spawnAsync({
-        cmd: "npm.cmd",
-        args: ["run-script", "build"],
-        cwd: "node_modules/pxt-core/skillmap",
-        shell: true
-    }).then(() => {
-        nodeutil.cpR("node_modules/pxt-core/skillmap/build", "docs/static/skillmap");
+function buildSkillMapAsync(parsed: commandParser.ParsedCommand) {
+    const skillmapRoot = "node_modules/pxt-core/skillmap";
+    // read pxtarget.json, save into 'pxtTargetBundle' global variable
+    let cfg = readLocalPxTarget();
+    nodeutil.writeFileSync(`${skillmapRoot}/public/target.js`, "// eslint-disable-next-line \n" + targetJsPrefix + JSON.stringify(cfg));
 
-        // patch paths to match updated folder structure
-        // patch <include src="file.html"> to <!-- @include file.html>
-        const fn = "docs/static/skillmap/index.html";
-        const f = fs.readFileSync(fn, "utf8");
-        const patched = f.replace(/\"\/static\//g, `"/static/skillmap/static/`).replace(/<include src="(\S+)">/gmi, "<!-- @include $1 -->");
-        fs.writeFileSync("docs/skillmap.html", patched, { encoding: "utf8" });
+    if (parsed.flags["serve"]) {
+        return nodeutil.spawnAsync({
+            cmd: "npm.cmd",
+            args: ["run-script", "start"],
+            cwd: skillmapRoot,
+            shell: true
+        })
+    } else {
+        return nodeutil.spawnAsync({
+            cmd: "npm.cmd",
+            args: ["run-script", "build"],
+            cwd: skillmapRoot,
+            shell: true
+        }).then(() => {
+            return rimrafAsync("docs/static/skillmap", {})
+        }).then(() => {
+            nodeutil.cpR(`${skillmapRoot}/build`, "docs/static/skillmap");
 
-        // remove old index.html
-        fs.unlinkSync(fn);
-        return Promise.resolve();
-    });
+            // patch paths to match updated folder structure
+            // patch <include src="file.html"> to <!-- @include file.html>
+            const fn = "docs/static/skillmap/index.html";
+            const f = fs.readFileSync(fn, "utf8");
+            const patched = f.replace(/href="\//g, `href="/static/skillmap/`)
+                .replace(/src="\//g, `src="/static/skillmap/`)
+                .replace(/<include src="(\S+)">/gmi, "<!-- @include $1 -->");
+            fs.writeFileSync("docs/skillmap.html", patched, { encoding: "utf8" });
+
+            // remove old index.html
+            fs.unlinkSync(fn);
+            return Promise.resolve();
+        })
+    }
 }
 
 function updateDefaultProjects(cfg: pxt.TargetBundle) {
@@ -6481,6 +6499,11 @@ ${pxt.crowdin.KEY_VARIABLE} - crowdin key
         aliases: ["skillmap"],
         advanced: true,
         help: "Builds the skill map webapp",
+        flags: {
+            serve: {
+                description: "Serve the skill map locally after building (npm start)"
+            }
+        }
     }, buildSkillMapAsync);
 
     advancedCommand("augmentdocs", "test markdown docs replacements", augmnetDocsAsync, "<temlate.md> <doc.md>");
